Favouriting feature nearly finished

+++ db/queries/favourites.js
@@ -19,12 +19,9 @@ const getFavouriteListings = (userID) => {
 
 
 const addToFavourites = (userID, listingID) => {
-  return db.query(`
-    INSERT INTO favourites (user_id, listing_id)
-    VALUES (${userID}, ${listingID});
-  `)
+  return db.query('INSERT INTO favourites (user_id, listing_id) VALUES ($1, $2) RETURNING *', [userID, listingID])
   .then (data => {
-    return data.rows;
+    return data.rows[0];
   })
   .catch(err => {
     return err.message;
@@ -33,13 +30,9 @@ const addToFavourites = (userID, listingID) => {
 
 
 const removeFromFavourites = (userID, listingID) => {
-  return db.query(`
-  DELETE FROM favourites
-  WHERE user_id = ${userID}
-  AND listing_id = ${listingID};
-  `)
+  return db.query('DELETE FROM favourites WHERE user_id = $1 AND listing_id = $2 RETURNING *', [userID, listingID])
   .then (data => {
-    return data.rows;
+    return data.rows[0];
   })
   .catch(err => {
     return err.message;
